enable saving and loading of phy_max for rapl cdev
Android Power HAL is reling on thermal daemon to provide
phy_max in case of throttling due to the low battery.
if system has been rebooted without thermal notifications
thermal daemon will not have correct phy_max available and
the system will be throttled on a very low levels. Save
calculated phy_max to a file and load it at every boot
if phy_max is not available from MSR.

Change-Id: Ia432874a4d5be35cc846c3c0336353d4fb677ebe
Signed-off-by: Radivoje Jovanovic <radivoje.jovanovic@intel.com>
Reviewed-on: https://android.intel.com/228889
Reviewed-by: cactus <cactus@intel.com>